Rework Signin Flow
Signin now works by 1 factor auth, either via phone or email.
The user is prompted to enter phone number or email address at login,
and, whether they're already signed in or not, they are sent a code
that will either log them in or create a new user for them.
Tres easy.